Restarting work - checkpointing before pruning some files from project

+++ framework/Source/GPUImageGammaFilter.m
@@ -5,7 +5,7 @@ NSString *const kGPUImageGammaFragmentShaderString = SHADER_STRING
     varying highp vec2 textureCoordinate;
 
     uniform sampler2D inputImage;
-    uniform lowp float gamma;
+    uniform highp float gamma;
 
     void main()
     {
